refactor(app): rename composition-root files to say what they hold
internal/app carried three files whose names all promised
initialization — app_initialization.go, initialization.go and
app_initialization_steps.go — plus a 37-line app_lifecycle.go that
held event-tap toggles, next to the real 600-line lifecycle.go. The
App struct itself hid in app_types.go while app.go held delegation
one-liners. The app_ prefix appeared on half the files and meant
nothing.

New names, all pure git renames:

  app_initialization.go        -> new.go            (New + phase driver)
  initialization.go            -> wiring.go         (adapter factories)
  initialization_platform_*.go -> wiring_platform_*.go
  app_initialization_steps.go  -> startup_phases.go (the numbered phases)
  app_lifecycle.go             -> eventtap_toggle.go
  app_types.go                 -> app.go            (the App struct)
  app.go                       -> mode_delegation.go
  app_config.go / app_getters.go / app_options.go
                               -> config.go / getters.go / options.go

The layering guardrail's composition-root allowlist and the docs that
referenced the old names are updated in the same change.
