+++ AGENTS.md
@@ -29,7 +29,7 @@ Neru follows a **Hexagonal Architecture (Ports and Adapters)**. All OS-specific 
 
 ### Finding the "Source of Truth"
 
-- **App Startup**: [app_initialization.go](internal/app/app_initialization.go)
+- **App Startup**: [new.go](internal/app/new.go)
 - **Navigation Logic**: [internal/app/modes/](internal/app/modes/)
 - **Coordinate Conversion**: [conversion.go](internal/domain/geometry/conversion.go)
 - **Error Definitions**: [errors.go](internal/derrors/errors.go)
