--- internal/app/app_types.go
@@ -1,117 +1,0 @@
-package app
-
-import (
-	"context"
-	"io"
-	"sync"
-
-	"go.uber.org/zap"
-
-	"github.com/y3owk1n/neru/internal/adapter/overlay/render/virtualpointer"
-	"github.com/y3owk1n/neru/internal/app/components"
-	"github.com/y3owk1n/neru/internal/app/ipcctrl"
-	"github.com/y3owk1n/neru/internal/app/keybinding"
-	"github.com/y3owk1n/neru/internal/app/modes"
-	"github.com/y3owk1n/neru/internal/app/sequence"
-	"github.com/y3owk1n/neru/internal/app/services"
-	"github.com/y3owk1n/neru/internal/app/services/modeindicator"
-	"github.com/y3owk1n/neru/internal/app/services/stickyindicator"
-	"github.com/y3owk1n/neru/internal/config"
-	"github.com/y3owk1n/neru/internal/domain"
-	"github.com/y3owk1n/neru/internal/domain/state"
-	"github.com/y3owk1n/neru/internal/ports"
-	"github.com/y3owk1n/neru/internal/ui"
-)
-
-// Mode is the current mode of the application.
-type Mode = domain.Mode
-
-// Mode constants from domain package.
-const (
-	ModeIdle   = domain.ModeIdle
-	ModeHints  = domain.ModeHints
-	ModeGrid   = domain.ModeGrid
-	ModeScroll = domain.ModeScroll
-)
-
-// SystrayComponent is the interface for systray functionality.
-type SystrayComponent interface {
-	OnReady()
-	OnExit()
-	Close()
-}
-
-// App is the main application instance containing all state and dependencies.
-type App struct {
-	ctx        context.Context //nolint:containedctx // Root context for all App operations
-	cancel     context.CancelFunc
-	config     *config.Config
-	ConfigPath string
-	logger     *zap.Logger
-
-	systemPort    ports.SystemPort
-	accessibility ports.AccessibilityPort
-
-	appState    *state.AppState
-	cursorState *state.CursorState
-
-	// Core services
-	overlayManager OverlayManager
-	hotkeyManager  HotkeyService
-	hotkeys        *keybinding.Binder
-	eventTap       ports.EventTapPort
-	textInput      ports.TextInputPort
-	keyFeed        ports.KeyFeedPort
-	ipcServer      ports.IPCPort
-	appWatcher     Watcher
-
-	modes *modes.Handler
-
-	// sequenceExecutor runs action sequences. It is built during initialization
-	// from the App's own components; see newSequenceExecutor.
-	sequenceExecutor *sequence.Executor
-
-	// axClient is stored so it can be closed during Cleanup.
-	// On Linux this resets AT-SPI accessibility status and releases the
-	// D-Bus connection; on other platforms it is a no-op.
-	axClient io.Closer
-
-	// Control channels
-	stopChan    chan struct{}
-	stopOnce    sync.Once
-	cleanupOnce sync.Once
-
-	// configMu serializes access to config-dependent component state between
-	// concurrent writers (theme change observer, IPC config reload, systray reload).
-	configMu sync.RWMutex
-
-	// New Architecture Services
-	hintService            *services.HintService
-	gridService            *services.GridService
-	actionService          *services.ActionService
-	scrollService          *services.ScrollService
-	modeIndicatorService   *modeindicator.Service
-	stickyIndicatorService *stickyindicator.Service
-	configService          *config.Service
-
-	// Feature components
-	hintsComponent           *components.HintsComponent
-	gridComponent            *components.GridComponent
-	scrollComponent          *components.ScrollComponent
-	modeIndicatorComponent   *components.ModeIndicatorComponent
-	stickyIndicatorComponent *components.StickyIndicatorComponent
-	recursiveGridComponent   *components.RecursiveGridComponent
-	virtualPointerOverlay    *virtualpointer.Overlay
-	systrayComponent         SystrayComponent
-
-	// Lifecycle management
-	gcCancel         context.CancelFunc
-	gcAggressiveMode bool
-
-	// State subscriptions
-	screenShareSubscriptionID uint64
-
-	renderer *ui.OverlayRenderer
-
-	ipcController *ipcctrl.Controller
-}
