+++ internal/app/startup_integration_darwin_test.go
