+++ internal/app/startup_phases.go
